Add files via upload
I designed 2 versions to give you the chance to pick between the 2 options. I did not go for any dark theme in these designs because I did not see any dark colour scheme in the branding repo we are to take images from. Looking forward to hearing your thoughts on these